refactor sensors, environment into own modules from lib
